+++ b2sdk/sync/sync.py
@@ -73,7 +73,7 @@ def zip_folders(folder_a, folder_b, reporter, policies_manager=DEFAULT_SCAN_MANA
             current_b = next_or_none(iter_b)
 
 
-def count_files(local_folder, reporter):
+def count_files(local_folder, reporter, policies_manager):
     """
     Count all of the files in a local folder.
 
@@ -82,9 +82,9 @@ def count_files(local_folder, reporter):
     """
     # Don't pass in a reporter to all_files.  Broken symlinks will be reported
     # during the next pass when the source and dest files are compared.
-    for _ in local_folder.all_files(None):
-        reporter.update_local(1)
-    reporter.end_local()
+    for _ in local_folder.all_files(None, policies_manager=policies_manager):
+        reporter.update_total(1)
+    reporter.end_total()
 
 
 @unique
@@ -96,6 +96,30 @@ class KeepOrDeleteMode(Enum):
 
 
 class Synchronizer(object):
+    """
+    Copies multiple "files" from source to destination.  Optionally
+    deletes or hides destination files that the source does not have.
+
+    The synchronizer can copy files:
+
+    - From a B2 bucket to a local destination.
+    - From a local source to a B2 bucket.
+    - From one B2 bucket to another.
+    - Between different folders in the same B2 bucket.
+      It will sync only the latest versions of files.
+
+    By default, the synchronizer:
+
+    - Fails when the specified source directory doesn't exist or is empty.
+      (see ``allow_empty_source`` argument)
+    - Fails when the source is newer.
+      (see ``newer_file_mode`` argument)
+    - Doesn't delete a file if it's present on the destination but not on the source.
+      (see ``keep_days_or_delete`` and ``keep_days`` arguments)
+    - Compares files based on modification time.
+      (see ``compare_version_mode`` and ``compare_threshold`` arguments)
+    """
+
     def __init__(
         self,
         max_workers,
@@ -171,11 +195,17 @@ class Synchronizer(object):
         :param int now_millis: current time in milliseconds
         :param b2sdk.sync.report.SyncReport,None reporter: progress reporter
         """
+        source_type = source_folder.folder_type()
+        dest_type = dest_folder.folder_type()
+
+        if source_type != 'b2' and dest_type != 'b2':
+            raise ValueError('Sync between two local folders is not supported!')
+
         # For downloads, make sure that the target directory is there.
-        if dest_folder.folder_type() == 'local' and not self.dry_run:
+        if dest_type == 'local' and not self.dry_run:
             dest_folder.ensure_present()
 
-        if source_folder.folder_type() == 'local' and not self.allow_empty_source:
+        if source_type == 'local' and not self.allow_empty_source:
             source_folder.ensure_non_empty()
 
         # Make an executor to count files and run all of the actions. This is
@@ -189,37 +219,26 @@ class Synchronizer(object):
         queue_limit = self.max_workers + 1000
         sync_executor = BoundedQueueExecutor(unbounded_executor, queue_limit=queue_limit)
 
-        # First, start the thread that counts the local files. That's the operation
-        # that should be fastest, and it provides scale for the progress reporting.
-        local_folder = None
-        if source_folder.folder_type() == 'local':
-            local_folder = source_folder
-        if dest_folder.folder_type() == 'local':
-            local_folder = dest_folder
-        if local_folder is None:
-            raise ValueError('neither folder is a local folder')
-        if reporter:
-            sync_executor.submit(count_files, local_folder, reporter)
+        if source_type == 'local' and reporter is not None:
+            # Start the thread that counts the local files. That's the operation
+            # that should be fastest, and it provides scale for the progress reporting.
+            sync_executor.submit(count_files, source_folder, reporter, self.policies_manager)
 
-        # Schedule each of the actions
-        bucket = None
-        if source_folder.folder_type() == 'b2':
-            bucket = source_folder.bucket
-        if dest_folder.folder_type() == 'b2':
-            bucket = dest_folder.bucket
-        if bucket is None:
-            raise ValueError('neither folder is a b2 folder')
-        total_files = 0
-        total_bytes = 0
+        # Bucket for scheduling actions.
+        # For bucket-to-bucket sync, the bucket for the API calls should be the destination.
+        action_bucket = None
+        if dest_type == 'b2':
+            action_bucket = dest_folder.bucket
+        elif source_type == 'b2':
+            action_bucket = source_folder.bucket
+
+        # Schedule each of the actions.
         for action in self.make_folder_sync_actions(
             source_folder, dest_folder, now_millis, reporter, self.policies_manager
         ):
-            logging.debug('scheduling action %s on bucket %s', action, bucket)
-            sync_executor.submit(action.run, bucket, reporter, self.dry_run)
-            total_files += 1
-            total_bytes += action.get_bytes()
-        if reporter:
-            reporter.end_compare(total_files, total_bytes)
+            logging.debug('scheduling action %s on bucket %s', action, action_bucket)
+            sync_executor.submit(action.run, action_bucket, reporter, self.dry_run)
+
         # Wait for everything to finish
         sync_executor.shutdown()
         if sync_executor.get_num_exceptions() != 0:
@@ -250,9 +269,11 @@ class Synchronizer(object):
         source_type = source_folder.folder_type()
         dest_type = dest_folder.folder_type()
         sync_type = '%s-to-%s' % (source_type, dest_type)
-        if (source_type, dest_type) not in [('b2', 'local'), ('local', 'b2')]:
-            raise NotImplementedError("Sync support only local-to-b2 and b2-to-local")
+        if source_type != 'b2' and dest_type != 'b2':
+            raise ValueError('Sync between two local folders is not supported!')
 
+        total_files = 0
+        total_bytes = 0
         for source_file, dest_file in zip_folders(
             source_folder,
             dest_folder,
@@ -264,12 +285,12 @@ class Synchronizer(object):
             elif dest_file is None:
                 logger.debug('determined that %s is not present on destination', source_file)
 
-            if source_folder.folder_type() == 'local':
-                if source_file is not None:
-                    reporter.update_compare(1)
-            else:
-                if dest_file is not None:
-                    reporter.update_compare(1)
+            if source_file is not None:
+                if source_type == 'b2':
+                    # For buckets we don't want to count files separately as it would require
+                    # more API calls. Instead, we count them when comparing.
+                    reporter.update_total(1)
+                reporter.update_compare(1)
 
             for action in self.make_file_sync_actions(
                 sync_type,
@@ -279,7 +300,16 @@ class Synchronizer(object):
                 dest_folder,
                 now_millis,
             ):
+                total_files += 1
+                total_bytes += action.get_bytes()
                 yield action
+
+        if reporter is not None:
+            if source_type == 'b2':
+                # For buckets we don't want to count files separately as it would require
+                # more API calls. Instead, we count them when comparing.
+                reporter.end_total()
+            reporter.end_compare(total_files, total_bytes)
 
     def make_file_sync_actions(
         self,