+++ test/unit/v0/apiver/apiver_deps.py
@@ -1,0 +1,11 @@
+######################################################################
+#
+# File: test/unit/v0/apiver/apiver_deps.py
+#
+# Copyright 2020 Backblaze Inc. All Rights Reserved.
+#
+# License https://www.backblaze.com/using_b2_code.html
+#
+######################################################################
+
+from b2sdk.v0 import *